Moved CreateTransaction from user folder to expense folder
